DR-391 Renaming service names
