+++ application/locations-data-manager/service.py
