+++ test/unit/healthcare-services-data-manager/test_service.py
@@ -1,4 +1,4 @@
-import oa_service
+import service
 import boto3
 
 from moto import mock_dynamodb
@@ -19,7 +19,7 @@ mock_revised_hospital_location = "RevisedLocation"
 def create_mock_dynamodb():
     "Create a mock implementation of the database table"
     dynamodb = boto3.resource("dynamodb")
-    table_name = oa_service.TABLE_NAME
+    table_name = service.TABLE_NAME
     table = dynamodb.create_table(
         TableName=table_name,
         KeySchema=[{"AttributeName": "id", "KeyType": "HASH"}],
@@ -44,8 +44,8 @@ def build_mock_data(id, hospital_name, hospital_location):
 def load_mock_data(data: map):
     "Load mock data to the table"
     dynamodb = boto3.resource("dynamodb")
-    table = dynamodb.Table(oa_service.TABLE_NAME)
-    table.put_item(Item=data, TableName=oa_service.TABLE_NAME)
+    table = dynamodb.Table(service.TABLE_NAME)
+    table.put_item(Item=data, TableName=service.TABLE_NAME)
 
 
 @mock_dynamodb
@@ -53,8 +53,8 @@ def test_get_record_by_id():
     "Test get_record_by_id method"
     table = create_mock_dynamodb()
     data = build_mock_data(mock_id, mock_hospital_name, mock_hospital_location)
-    table.put_item(Item=data, TableName=oa_service.TABLE_NAME)
-    response = oa_service.get_record_by_id(mock_id)
+    table.put_item(Item=data, TableName=service.TABLE_NAME)
+    response = service.get_record_by_id(mock_id)
     assert response["Item"]["id"] == mock_id
     assert response["Item"]["HospitalName"] == mock_hospital_name
     assert response["Item"]["HospitalLocation"] == mock_hospital_location
@@ -69,7 +69,7 @@ def test_add_record():
     write_data = build_mock_data(
         mock_post_id, mock_post_hospital_name, mock_post_hospital_location
     )
-    oa_service.add_record(write_data)
+    service.add_record(write_data)
     response = table.get_item(Key={"id": mock_post_id})
     assert response["Item"]["id"] == mock_post_id
     assert response["Item"]["HospitalName"] == mock_post_hospital_name
@@ -81,12 +81,12 @@ def test_update_record():
     "Test update record method - eg used by PUT"
     table = create_mock_dynamodb()
     data = build_mock_data(mock_id, mock_hospital_name, mock_hospital_location)
-    table.put_item(Item=data, TableName=oa_service.TABLE_NAME)
+    table.put_item(Item=data, TableName=service.TABLE_NAME)
     response = table.get_item(Key={"id": mock_id})
     assert response["Item"]["id"] == mock_id
     assert response["Item"]["HospitalName"] == mock_hospital_name
     assert response["Item"]["HospitalLocation"] == mock_hospital_location
-    response = oa_service.update_record(
+    response = service.update_record(
         mock_id, mock_revised_hospital_location, mock_revised_hospital_name
     )
     assert response["Attributes"]["HospitalName"] == mock_revised_hospital_name
@@ -98,9 +98,9 @@ def test_delete_record_by_id():
     "Test delete_record method first adding and then checking it exists"
     table = create_mock_dynamodb()
     data = build_mock_data(mock_id, mock_hospital_name, mock_hospital_location)
-    table.put_item(Item=data, TableName=oa_service.TABLE_NAME)
-    response = oa_service.get_record_by_id(mock_id)
+    table.put_item(Item=data, TableName=service.TABLE_NAME)
+    response = service.get_record_by_id(mock_id)
     assert response["Item"]["id"] == mock_id
-    oa_service.delete_record(mock_id)
+    service.delete_record(mock_id)
     response = table.get_item(Key={"id": mock_id})
     assert ("Item" in response) is False
